requirements.txt
asgiref==3.7.2
Django==5.0.1
django-simple-history==3.4.0
django-environ==0.11.2
gunicorn==21.2.0
packaging==23.2
psycopg2-binary==2.9.9
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0